Move healthcheck out of plugin into core
