--- plugins/API/HealthCheck/Checks/DashboardAvailableHealthCheck.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Piwik\Plugins\API\HealthCheck\Checks;
-
-use Piwik\Config\GeneralConfig;
-
-final class DashboardAvailableHealthCheck implements HealthCheckInterface
-{
-    public function getName(): string
-    {
-        return 'dashboardAvailable';
-    }
-
-    public function test(): bool
-    {
-        return GeneralConfig::getConfigValue('maintenance_mode') !== 1;
-    }
-}
