+++ tests/PHPUnit/Unit/HealthCheck/HealthCheck/Responses/HealthCheckResponseTest.php
@@ -1,14 +1,13 @@
 <?php
 
-namespace Piwik\Plugins\API\tests\Unit\HealthCheck\Responses;
+namespace HealthCheck\HealthCheck\Responses;
 
 use PHPUnit\Framework\TestCase;
-use Piwik\Plugins\API\HealthCheck\Responses\HealthCheckResponse;
-use Piwik\Plugins\API\HealthCheck\Responses\HealthCheckSingleResponse;
+use Piwik\HealthCheck\HealthCheck\Responses\HealthCheckResponse;
+use Piwik\HealthCheck\HealthCheck\Responses\HealthCheckSingleResponse;
 
 /**
- * @group Plugin
- * @group API
+ * @group Core
  * @group HealthCheck
  * @group HealthCheckResponse
  */
